requirements.txt
matplotlib==3.4.3
librosa==0.8.1
numpy==1.21.2
python-dotenv==0.17.1
mido==1.2
midi2audio==0.1.1
tqdm==4.62.2
torch==1.8.2
pytube==11.0.1
moviepy==1.0.3
midiutil==1.2.1